requirements.txt
httpx==0.25.1
beautifulsoup4==4.12.3
html5lib==1.1
feedparser==6.0.10
python-dotenv==1.0.1
jinja2==3.1.4
pydantic==2.8.2
openai==1.44.0
supabase==2.4.0
PyYAML==6.0.2
python-dateutil==2.9.0.post0
